style: suppress detekt warning about missing kdoc
It looks like detekt is having an issue noticing there is documentation for
the new fun interface for Kotlin, so for now suppress the warning about there
being no docs for it, as the docs do exist. Unable to put the suppression
directly on the interface as it doesn't seem to take notice of it.