refactor: change the old java LogHandler to Kotlin (#111)
* refactor: change the old java LogHandler to Kotlin

The LogHandler was written in Java so we could do SAM conversion
to make it a bit nicer to use from Java and Kotlin, now that
Kotlin 1.4.0 is out we can use a Kotlin interface and still get
the nice SAM conversion while keeping full Kotlin code.

* style: suppress detekt warning about missing kdoc

It looks like detekt is having an issue noticing there is documentation for
the new fun interface for Kotlin, so for now suppress the warning about there
being no docs for it, as the docs do exist. Unable to put the suppression
directly on the interface as it doesn't seem to take notice of it.